// src/services/models/Course.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Course = {
  adminId?: number;
  categoryId?: number;
  courseType?: number;
  coverImage?: string;
  createTime?: string;
  description?: string;
  difficulty?: number;
  id?: number;
  isDelete?: number;
  objectives?: string;
  originalPrice?: number;
  price?: number;
  ratingCount?: number;
  ratingScore?: number;
  requirements?: string;
  status?: number;
  studentCount?: number;
  subtitle?: string;
  tags?: string;
  targetAudience?: string;
  teacherId?: number;
  title?: string;
  totalChapters?: number;
  totalDuration?: number;
  totalSections?: number;
  updateTime?: string;
};